from int to double

+++ dz5/Model/DivModel.java
@@ -7,17 +7,17 @@ public class DivModel  extends CalcModel{
     }
     // do_it
     @Override
-    public int result() {
+    public double result() {
         return x / y;
     }
 
     @Override
-    public void setX(int value) {
+    public void setX(double value) {
         super.x = value;
     }
 
     @Override
-    public void setY(int value) {
+    public void setY(double value) {
         super.y = value;
     }
