+++ dz5/Model/Model.java
@@ -1,9 +1,9 @@
 package dz5.Model;
 
 public interface Model {
-    int result();
+    double result();
 
-    void setX(int value);
+    void setX(double value);
 
-    void setY(int value);
+    void setY(double value);
 }
